test(jabs-core): move JabsPoseVersion tests to own file
Task 1 appended tests to test_enums.py, an untracked WIP file not on main
(and carrying a failing test). Move the JabsPoseVersion tests into their own
test_pose_version.py and stop tracking test_enums.py so the PR ships no
pre-existing WIP/failing tests.
